Splitting oemof generatremodel and model
